requirements.txt
asgiref==3.2.7
atomicwrites==1.4.0
colorama==0.4.4
Django==3.0.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.7.0
flake8==3.9.2
flake8-quotes==3.3.0
gunicorn==20.0.4
importlib-metadata==4.8.1
isort==5.8.0
mccabe==0.6.1
pluggy==0.13.1
psycopg2-binary==2.8.5
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
python-dotenv==0.18.0
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
zipp==3.5.0
